requirements.txt
tk
pygame
pydub
librosa
matplotlib
pillow
pandas
fpdf
speechrecognition
pyaudio
mysql-connector-python
